+++ packages/@expo/cli/src/export/saveAssets.ts
@@ -52,6 +52,8 @@ export type ExportAssetDescriptor = {
   apiRouteId?: string;
   /** Expo Router route path for formatting the RSC output. */
   rscId?: string;
+  /** Expo Router route path for formatting the loader module output. */
+  loaderId?: string;
   /** A key for grouping together output files by server- or client-side. */
   targetDomain?: 'server' | 'client';
 };
@@ -76,6 +78,7 @@ export async function persistMetroFilesAsync(files: ExportAssetMap, outputDir: s
   const middlewareEntries: [string, ExportAssetDescriptor][] = [];
   const routeEntries: [string, ExportAssetDescriptor][] = [];
   const rscEntries: [string, ExportAssetDescriptor][] = [];
+  const loaderEntries: [string, ExportAssetDescriptor][] = [];
   const remainingEntries: [string, ExportAssetDescriptor][] = [];
 
   let hasServerOutput = false;
@@ -86,6 +89,7 @@ export async function persistMetroFilesAsync(files: ExportAssetMap, outputDir: s
     else if (asset[1].middlewareId != null) middlewareEntries.push(asset);
     else if (asset[1].apiRouteId != null) apiRouteEntries.push(asset);
     else if (asset[1].rscId != null) rscEntries.push(asset);
+    else if (asset[1].loaderId != null) loaderEntries.push(asset);
     else remainingEntries.push(asset);
   }
 
@@ -257,6 +261,18 @@ export async function persistMetroFilesAsync(files: ExportAssetMap, outputDir: s
         sizeStr(assets.contents),
         hasSourceMap ? chalk.gray(`(source map ${sizeStr(hasSourceMap[1].contents)})`) : ''
       );
+    }
+  }
+
+  if (loaderEntries.length) {
+    Log.log('');
+    Log.log(chalk.bold`${BLT} Loader modules (${loaderEntries.length}):`);
+
+    for (const [loaderFilename, assets] of loaderEntries.sort(
+      (a, b) => a[0].length - b[0].length
+    )) {
+      const id = assets.loaderId!;
+      Log.log(id === '/' ? '/ ' + chalk.gray('(index)') : id, sizeStr(assets.contents));
     }
   }